pyproject.toml
[project]
name = "mr-helpmate-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "chromadb>=0.5.0",
    "google-generativeai>=0.8.0",
    "gradio>=4.0.0",
    "langchain[google-genai]>=0.3.27",
    "langgraph>=0.2.0",
    "numpy>=1.24.0",
    "openai>=2.2.0",
    "pandas>=2.0.0",
    "pdfplumber>=0.11.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "ruff>=0.13.3",
    "sentence-transformers>=2.2.2",
    "tqdm>=4.66.0",
    "transformers>=4.44.0",
]
